+++ nilearn/image/tests/test_image.py
@@ -7,6 +7,8 @@ from nose import SkipTest
 import platform
 import os
 import sys
+import tempfile
+
 import nibabel
 from nibabel import Nifti1Image
 import numpy as np
@@ -21,7 +23,7 @@ from nilearn.image import concat_imgs
 from nilearn._utils import testing, niimg_conversions, data_gen
 from nilearn.image import new_img_like
 from nilearn.image import threshold_img
-from nilearn.image import iter_img
+from nilearn.image import iter_img, index_img
 from nilearn.image import math_img
 from nilearn.image import largest_connected_component_img
 from nilearn.image import get_data
@@ -35,6 +37,25 @@ X64 = (platform.architecture()[0] == '64bit')
 
 currdir = os.path.dirname(os.path.abspath(__file__))
 datadir = os.path.join(currdir, 'data')
+
+
+def test_get_data():
+    img, *_ = data_gen.generate_fake_fmri(shape=(10, 11, 12))
+    data = get_data(img)
+    assert data.shape == img.shape
+    assert data is img._data_cache
+    mask_img = new_img_like(img, data > 0)
+    data = get_data(mask_img)
+    assert data.dtype == np.dtype('int8')
+    img_3d = index_img(img, 0)
+    with tempfile.TemporaryDirectory() as tempdir:
+        filename = os.path.join(tempdir, 'img_{}.nii.gz')
+        img_3d.to_filename(filename.format('a'))
+        img_3d.to_filename(filename.format('b'))
+        data = get_data(filename.format('a'))
+        assert len(data.shape) == 3
+        data = get_data(filename.format('*'))
+        assert len(data.shape) == 4
 
 
 def test_high_variance_confounds():